PERHITUNGAN/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
hijri-converter==2.3.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.4.0
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
secure-smtplib==0.1.1
six==1.16.0
tzdata==2023.3
Werkzeug==3.0.1
WTForms==3.1.2
